fix bosscontroller and effect blood takedame has direction

+++ _Scrips/Player/Player Behaviour/HeliSlamState.cs
@@ -4,7 +4,7 @@ public class HeliSlamState : PlayerState
 {
     private float slamSpeed = 60;
     private bool hasExploded = false;
-    private float damage = 50; // Dame chiêu
+    private int damage = 50; // Dame chiêu
     private float attackRange = 2f; // Tầm xa chiêu
 
     public HeliSlamState(PlayerController player) : base(player) { }
