+++ _Scrips/Player/Player Behaviour/QuickAttackState.cs
@@ -4,7 +4,7 @@ public class QuickAttackState : PlayerState
 {
     private float attackStartTime;
     private float attackDuration = 0.33f; // Thời gian hồi chiêu
-    private float damage = 10; // Dame chiêu
+    private int damage = 10; // Dame chiêu
     private float attackRange = 0.8f; // Tầm xa chiêu
 
     public QuickAttackState(PlayerController player) : base(player) { }
